refactor(publication-header): move to publication_header package
