+++ lib/views/components/content/publication_header/publication_header.dart
