+++ lib/views/pages/post/components/complete_post.dart
@@ -1,7 +1,7 @@
 import "package:flutter/material.dart";
 import "package:proxima/models/ui/post_details.dart";
 import "package:proxima/views/components/content/post_votes.dart";
-import "package:proxima/views/components/content/publication_header.dart";
+import "package:proxima/views/components/content/publication_header/publication_header.dart";
 
 class CompletePost extends StatelessWidget {
   static const postTitleKey = Key("postTitle");
